feat(docs): Use preprocessor syntax for including versions
